requirements.txt
databricks-cli>=0.18.0
